+++ bedboss/refgenome_validator/genome_model.py
@@ -1,6 +1,10 @@
+from typing import Union, Dict
+
+
 class GenomeModel:
     """
         Initialize genome model
+
     A class representing a reference genome. You feed it a reference genome. It retrieves chrom sizes (from refgenie),
     and then it provides some helper functions, intended for use with reference genome validation.
 
@@ -9,21 +13,25 @@ class GenomeModel:
     def __init__(
         self,
         genome_alias: str,
-        chrom_sizes_file: str,
+        chrom_sizes_file: Union[str, Dict[str, int]],
+        genome_digest: Union[str, None] = None,
         # common_aliases: Optional[List] = None,
-        # refgenomeconf: Optional[refgenconf.refgenconf.RefGenConf] = None,
         # exclude_ranges_names: Optional[List] = None,
     ):
         self._genome_alias = genome_alias
         self.chrom_sizes_file = chrom_sizes_file
+        self._genome_digest = genome_digest
         self._chrom_sizes = self.get_chrom_sizes()
         # self.common_aliases = common_aliases  # What are the other names for the other this reference genomes
-        # self.rgc = refgenomeconf
         # self.excluded_ranges_names = exclude_ranges_names  # Which bed file digests from the excluded ranges are associated with this reference genome?
 
     @property
     def genome_alias(self):
-        return self._genome_alias
+        return self._genome_digest  # TODO: update everything
+
+    @property
+    def genome_digest(self):
+        return self._genome_digest
 
     @property
     def chrom_sizes(self):
@@ -36,22 +44,18 @@ class GenomeModel:
         :return dict: dictionary containing chroms(keys) and lengths(values)
         """
 
-        # if self.rgc:
-        #     chrom_sizes_path = self.rgc.seek(
-        #         genome_name=self.genome_alias,
-        #         asset_name="fasta",
-        #         tag_name="default",
-        #         seek_key="chrom_sizes",
-        #     )
+        if isinstance(self.chrom_sizes_file, str):
+            chrom_sizes_path = self.chrom_sizes_file
 
-        chrom_sizes_path = self.chrom_sizes_file
+            chrom_sizes = {}
 
-        chrom_sizes = {}
-
-        with open(chrom_sizes_path, "r") as f:
-            for line in f:
-                chrom, size = line.strip().split("\t")
-                chrom_sizes[chrom] = int(size)
+            with open(chrom_sizes_path, "r") as f:
+                for line in f:
+                    chrom, size = line.strip().split("\t")
+                    chrom_sizes[chrom] = int(size)
+        else:
+            # TODO: needs to be validated
+            chrom_sizes = self.chrom_sizes_file
 
         return chrom_sizes